Fixed JPQL query related bugs
* Fixed JPQL query bugs in car, garage and maintenance repositories
* Modified report DTOs to work with JPQL query projections
* Modified car, garage and maintenance mappers to set null instead of 0
for the id when creation entities
* Modified car mapper to avoid null pointer reference while mapping
without provided garage ids
* Modified car service to avoid null pointer reference while updating
car if garage ids aren't provided
* Modified gitignore to not include database files

+++ carmanagement/src/main/java/pu/fmi/wsp/hw/carmanagement/model/dto/report/GarageDailyAvailabilityReportDTO.java
@@ -10,10 +10,10 @@ public class GarageDailyAvailabilityReportDTO {
 	public GarageDailyAvailabilityReportDTO() {
 	}
 
-	public GarageDailyAvailabilityReportDTO(LocalDate date, int requests, int availableCapacity) {
+	public GarageDailyAvailabilityReportDTO(LocalDate date, long requests, long availableCapacity) {
 		this.date = date;
-		this.requests = requests;
-		this.availableCapacity = availableCapacity;
+		this.requests = (int)requests;
+		this.availableCapacity = (int)availableCapacity;
 	}
 
 	public LocalDate getDate() {
